login system

+++ naidisprojekt/Models/Product.cs
@@ -9,10 +9,9 @@ namespace naidisprojekt.Models
     public class Product
     {
         public int Id { get; set; }
-        public string 
-            Name { get; set; }    
+        public string Name { get; set; }    
         public string? Description { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int CategoryId {  get; set; }  
         public string? ImageSource { get; set; }
     }
